Rename PrepareRender to PrepareResources, and add new flags.
Added prRender (you need to explicitly request to prepare render now). Fixed every program to include prRender now.

Added prSpatial, to allow you to prepare octrees --- otherwise they may be prepared on demand, esp. the TVRMLShape.OctreeTriangles in case of ssDynamicCollisions may be prepared in the middle of walking.

For now, castle uses prSpatial. For view3dscene, not preparing prSpatial (deferring it to when it's actually needed) may be good.