Removed examples/images/dds_remove_small_mipmaps.
This was a tool to remove from DDS mipmaps the levels with one of the sizes < 4. It workarounds bug in older version of GIMP-DDS, it generated invalid images for smaller mipmaps. The bug is fixed in GIMP-DDS >= 2.0.9. See http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=564111 .

Since the fixed version of GIMP-DDS is available since a long time (and the tool didn't really serve any other purpose), it's removed now.