Add DoProgress parameter to PrepareCore, make progress optional.
This is important if Prepare happens in the middle of another progress, e.g. when you forgot to add initial creature to <resources> in level.

Remove AnimationName param and related log messages, not terribly useful and were complicating code.

GameLevel merely warns when not CreatureKind.Prepared. It will be handled 100% Ok now.