Rename TCamera.NavigationInput to simpler TCamera.Input, along with related type name.
Rename ciKeyboard (previously niKeyboard) to ciNormal. That's because all Input_Xxx shortcuts describe both key and mouse inputs, and they affect both key/mouse-button/mouse-wheel holding (when Idle does something continous when key is down) or key/mouse-button/mouse-wheel clicking (when EventDown does something once when input is pressed).
So, the new name "ciNormal" is quite poor and non-descriptive, but that's because it really toggles various input devices and reactions.