glTexImage2DMultisample troubles resolved. We can actually make them work Ok. Further commits will add fixes to GLSL code all around, for now we enable them in code by default.
The sampler2DMS should always be addressed with texelFetch, that takes integer pixel positions. So actually this is quite close to texture rectangles, and is ideal for our purposes.

To have behavior uniform between texture rects (no AA) and texture MS (AA), we will need to link our simple functions anyway, to hide differences.

Remove TGLRenderToTexture.EnableFramebuffer property: this was only for a temporary hack.