Make triangulation working really Ok for duplicate vertexes. Duplicates are a problem, as they may cause non-empty triangle to be considered empty (because some other vertexes lie on the border of ear triangle, and are then ignored).
Removing successive duplicates (as done previously) was not a full solution, our visualization shows it. That is because duplicates need not occur on successive position to cause problems, they may be separated on both sides (by non-zero-area triangles). Also, any colinear situations are problematic, not just with 2 equal vertexes. The previous duplication removal worked for polygon_3_5 because it accidentally also incremented P0 by one step, which meant that vertexes 3 and 6 could be removed (while in general we have 1-3-5-6 sequence, and 3 and 6 are equal).

Current approach actually looks what happens when a vertex is on the border of a considered ear. Depending on the direction in which it is pulled, we may either ignore or take into account this collision.

Also, more info at VISUALIZE_TRIANGULATION.