Remake TLevel inside castle. The goal is to make it more componentized, and see if we can salvage parts of it for generic engine unit (and so, useful also for castle2 and other games).
This commit is a forced separation of scene manager and level logic. Previously TLevel was our TCastleSceneManager, and it had descendants that customize it for particular level. This means that loading new level equals destroying + creating new scene manager, which is a little unhandy, as we then release all prepared 3D stuff, and lose all scene manager configuration. Which isn't really a problem now, but it goes against our preferred scene manager usage.

This commit makes new TLevel class as T3D descendant, that is more independent from scene manager. We need this, as some level-specific logic really needs to exist and change during level play, so we can't just shove it inside TLevelAvailable. Scene manager is in separate TGameSceneManager.