Looks like cube maps can work on affected Intel GPUs --- but only with shader rendering.
Introduce BuggyFixedFunctionCubeMap, forces shader rendering when 1. we're on Intel GPU, 2. the GPU is capable of shaders (so it's not old Intel GPU...), 3. we're rendering shape with a cube map (the other shapes will remain using fixed-function pipeline, so all will be hopefully fast).

Remove BuggyGenerateCubeMap, since we found a way to keep them working :)