Improve "The Castle" to work with any fullscreen size.
The main change is that castle.lpr sets FullScreen:=true at the beginning, and works in fullscreen even if screen resize failed. Useful for CastleWindow GTK backend (that cannot resize the screen). This also simplifies the code a little, and allows to accept standard --geometry and such options.

Other changes improve behavior of the game in window sizes other than 800x600:
* Over background: Position on-screen menu in top-left corner.
* Over game: Position on-screen menu in exact center, and draw fade rect in center. Also, draw fade rect simpler.
* Background image scaling improve to riBilinear.
* credits.wrl size fits better in common resolutions.
* Manpages and WWW docs updated to refect new command-line options and behavior.